Add extraction schema id in custom deserializer. redesign columns selector. Fix review issues.
(cherry picked from commit 970cd2bb5bee2a84b389fa844c2ba9fd2c286e7d)
